COMMERCE-960 Move IdUtils to Core
