+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/util/v1_0/UserHelper.java
@@ -14,11 +14,11 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.util.v1_0;
 
-import com.liferay.commerce.openapi.admin.internal.util.IdUtils;
 import com.liferay.commerce.openapi.admin.internal.util.v1_0.DTOUtils;
 import com.liferay.commerce.openapi.admin.model.v1_0.UserDTO;
 import com.liferay.commerce.openapi.core.context.Pagination;
 import com.liferay.commerce.openapi.core.model.CollectionDTO;
+import com.liferay.commerce.openapi.core.util.IdUtils;
 import com.liferay.external.reference.service.ERUserLocalService;
 import com.liferay.portal.kernel.exception.NoSuchUserException;
 import com.liferay.portal.kernel.exception.PortalException;
