+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/v2_0/WebSiteResourceImpl.java
@@ -14,13 +14,13 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.v2_0;
 
-import com.liferay.commerce.openapi.admin.internal.util.IdUtils;
 import com.liferay.commerce.openapi.admin.internal.util.v2_0.DTOUtils;
 import com.liferay.commerce.openapi.admin.model.v2_0.WebSiteDTO;
 import com.liferay.commerce.openapi.admin.resource.v2_0.WebSiteResource;
 import com.liferay.commerce.openapi.core.context.Language;
 import com.liferay.commerce.openapi.core.context.Pagination;
 import com.liferay.commerce.openapi.core.model.CollectionDTO;
+import com.liferay.commerce.openapi.core.util.IdUtils;
 import com.liferay.oauth2.provider.scope.RequiresScope;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Company;
